update member types of wager response.

+++ app/wager/request.go
@@ -1,8 +1,8 @@
 package wager
 
 type CreateWagerRequest struct {
-	TotalWagerValue   uint `json:"total_wager_value"`
-	Odds              uint `json:"odds"`
-	SellingPercentage uint `json:"selling_percentage"`
-	SellingPrice      uint `json:"selling_price"`
+	TotalWagerValue   uint32  `json:"total_wager_value"`
+	Odds              uint32  `json:"odds"`
+	SellingPercentage uint32  `json:"selling_percentage"`
+	SellingPrice      float32 `json:"selling_price"`
 }
